feat: 로또 단말기 기능 추가
* 로또 티켓 자동 생성 기능을 로또 티켓에서 로또 단말기로 이동
